Move beans to two new packages
