+++ dev/com.ibm.ws.jsf.2.2_fat/test-applications/JSF22StatelessView.war/src/com/ibm/ws/jsf22/fat/statelessview/beans/ViewScopedCDIBean.java
@@ -11,7 +11,7 @@
 /**
  * A simple managed bean that will be used to test
  * a view scoped managed bean.
- * 
+ *
  * @author Bill Lucy
  *
  */
